old-tests/submission/Microsoft/video/video_015.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: 'canPlayType' method returns the empty string for 'application/octet-stream'.</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#dom-navigator-canplaytype" />
  <meta name="assert" content="'canPlayType' method returns the empty string for 'application/octet-stream'." />
</head>
<body>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <div id=log></div>
  
    <script type="text/javascript">

        test(function() { assert_equals(document.createElement("video").canPlayType("application/octet-stream"), "")}, "'HTMLAudioElement' inherits 'playbackRate' from 'HTMLMediaElement' and supports getting its default value");
  
    </script>
</div>
</body>
</html>
